backend/requirements/base.txt
Django>=5.0
djangorestframework>=3.14.0
django-allauth>=0.58.2
Pillow>=10.1.0
django-cors-headers>=4.3.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
razorpay>=1.4.1
celery>=5.3.4
redis>=5.0.1
django-filter>=23.3
django-crispy-forms>=2.1
pandas>=2.1.1
openpyxl>=3.1.2